Move mocks to tests
