--- projects/rocprofiler-systems/source/lib/backends/amd_smi/backend.cpp
@@ -1,6 +1,0 @@
-// Copyright (c) Advanced Micro Devices, Inc.
-// SPDX-License-Identifier: MIT
-
-#include "backends/amd_smi/backend.hpp"
-#include "backends/amd_smi/gpu_backend.hpp"
-#include "backends/amd_smi/nic_backend.hpp"
