--- projects/rocprofiler-systems/source/lib/backends/procfs/backend.cpp
@@ -1,4 +1,0 @@
-// Copyright (c) Advanced Micro Devices, Inc.
-// SPDX-License-Identifier: MIT
-
-#include "backends/procfs/backend.hpp"
