+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/cpu/tests/CMakeLists.txt
@@ -17,7 +17,7 @@ target_link_libraries(
         rocprofiler-systems-logger
         rocprofiler-systems-common-library
         rocprofiler-systems-user-library
-        rocprofiler-systems::rocprofiler-systems-procfs-backend
+        rocprofiler-systems::rocprofiler-systems-headers
 )
 
 target_include_directories(
