+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/gpu/tests/CMakeLists.txt
@@ -9,7 +9,8 @@ target_link_libraries(
         rocprofiler-systems-googletest-library
         rocprofiler-systems-logger
         rocprofiler-systems::rocprofiler-systems-headers
-        rocprofiler-systems::rocprofiler-systems-amd-smi-backend
+        rocprofiler-systems::rocprofiler-systems-rocm
+        rocprofiler-systems::rocprofiler-systems-compile-definitions
 )
 
 if(ROCPROFSYS_BUILD_AINIC)
