+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/gpu_perf_counter/tests/CMakeLists.txt
@@ -17,6 +17,5 @@ target_link_libraries(
         rocprofiler-systems-logger
         rocprofiler-systems::rocprofiler-systems-headers
         rocprofiler-systems::rocprofiler-systems-rocm
-        rocprofiler-systems::rocprofiler-systems-rocprofiler-sdk-backend
         rocprofiler-systems::rocprofiler-systems-timemory
 )
