+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/nic/tests/CMakeLists.txt
@@ -8,7 +8,9 @@ target_link_libraries(
     PRIVATE
         rocprofiler-systems-googletest-library
         rocprofiler-systems-logger
-        rocprofiler-systems::rocprofiler-systems-amd-smi-backend
+        rocprofiler-systems::rocprofiler-systems-headers
+        rocprofiler-systems::rocprofiler-systems-rocm
+        rocprofiler-systems::rocprofiler-systems-compile-definitions
 )
 
 if(ROCPROFSYS_BUILD_AINIC)
