+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/sampler.cpp
@@ -39,7 +39,6 @@
 
 #include "logger/debug.hpp"
 
-#include <amd_smi/amdsmi.h>
 #include <timemory/backends/threading.hpp>
 #include <timemory/components/timing/backends.hpp>
 #include <timemory/mpl/type_traits.hpp>
